@@ -46,13 +46,13 @@
 		"selenium-webdriver": "4.4.0",
 		"shelljs": "^0.8.5",
 		"ts-node": "^10.9.1",
-		"typescript": "3.9.9",
+		"typescript": "4.9.4",
 		"vscode-extension-tester-locators": "3.1.0",
 		"yaml": "^2.2.2"
 	},
 	"dependencies": {
 		"@eclipse-che/api": "latest",
-		"inversify": "5.0.1",
+		"inversify": "6.0.1",
 		"reflect-metadata": "0.1.13"
 	},
 	"resolutions": {
