@@ -18,7 +18,13 @@ import { OAUTH_CONSTANTS } from '../../../constants/OAUTH_CONSTANTS';
 
 @injectable()
 export class OcpLoginPage {
-	private static readonly LOGIN_PAGE_OPENSHIFT_XPATH: string = '//*[contains(text(), "Welcome")]';
+	private static readonly LOGIN_PAGE_WELCOME_MESSAGE: By = By.xpath('//*[contains(text(), "Welcome")]');
+	private static readonly LOGIN_BUTTON: By = By.css('button[type=submit]');
+	private static readonly LOGIN_PROVIDER_BUTTON: By = By.xpath(`//a[text()="${OAUTH_CONSTANTS.TS_OCP_LOGIN_PAGE_PROVIDER_TITLE}"]`);
+	private static readonly AUTHORIZE_OPENSHIFT_ACCESS_HEADER: By = By.xpath('//h1[text()="Authorize Access"]');
+	private static readonly APPROVE_ACCESS_BUTTON: By = By.css('input[name="approve"]');
+	private static readonly USERNAME_INPUT: By = By.id('inputUsername');
+	private static readonly PASSWORD_INPUT: By = By.id('inputPassword');
 
 	constructor(
 		@inject(CLASSES.DriverHelper)
@@ -28,72 +34,54 @@ export class OcpLoginPage {
 	async waitOpenShiftLoginWelcomePage(): Promise<void> {
 		Logger.debug();
 
-		await this.driverHelper.waitVisibility(
-			By.xpath(OcpLoginPage.LOGIN_PAGE_OPENSHIFT_XPATH),
-			TIMEOUT_CONSTANTS.TS_SELENIUM_LOAD_PAGE_TIMEOUT
-		);
+		await this.driverHelper.waitVisibility(OcpLoginPage.LOGIN_PAGE_WELCOME_MESSAGE, TIMEOUT_CONSTANTS.TS_SELENIUM_LOAD_PAGE_TIMEOUT);
 	}
 
-	async clickOnLoginProviderTitle(): Promise<void> {
+	async waitAndClickOnLoginProviderTitle(): Promise<void> {
 		Logger.debug();
 
-		const loginProviderTitleLocator: By = By.xpath(`//a[text()="${OAUTH_CONSTANTS.TS_OCP_LOGIN_PAGE_PROVIDER_TITLE}"]`);
-		await this.driverHelper.waitAndClick(loginProviderTitleLocator, TIMEOUT_CONSTANTS.TS_SELENIUM_WAIT_FOR_URL);
+		await this.driverHelper.waitAndClick(OcpLoginPage.LOGIN_PROVIDER_BUTTON, TIMEOUT_CONSTANTS.TS_SELENIUM_WAIT_FOR_URL);
 	}
 
 	async isIdentityProviderLinkVisible(): Promise<boolean> {
 		Logger.debug();
 
-		const loginWithHtpaswdLocator: By = By.xpath(`//a[text()="${OAUTH_CONSTANTS.TS_OCP_LOGIN_PAGE_PROVIDER_TITLE}"]`);
-		return await this.driverHelper.waitVisibilityBoolean(loginWithHtpaswdLocator, 3, 5000);
+		return await this.driverHelper.waitVisibilityBoolean(OcpLoginPage.LOGIN_PROVIDER_BUTTON, 3, 5000);
 	}
 
 	async isAuthorizeOpenShiftIdentityProviderPageVisible(): Promise<boolean> {
 		Logger.debug();
 
-		const authorizeOpenshiftIdentityProviderPageLocator: By = By.xpath('//h1[text()="Authorize Access"]');
-		return await this.driverHelper.isVisible(authorizeOpenshiftIdentityProviderPageLocator);
-	}
-
-	async waitAuthorizeOpenShiftIdentityProviderPage(): Promise<void> {
-		Logger.debug();
-
-		const authorizeOpenshiftIdentityProviderPageLocator: By = By.xpath('//h1[text()="Authorize Access"]');
-		await this.driverHelper.waitVisibility(
-			authorizeOpenshiftIdentityProviderPageLocator,
-			TIMEOUT_CONSTANTS.TS_SELENIUM_LOAD_PAGE_TIMEOUT
-		);
+		return await this.driverHelper.isVisible(OcpLoginPage.AUTHORIZE_OPENSHIFT_ACCESS_HEADER);
 	}
 
 	async clickOnApproveAuthorizeAccessButton(): Promise<void> {
 		Logger.debug();
 
-		const approveAuthorizeAccessOcpLocator: By = By.css('input[name="approve"]');
-		await this.driverHelper.waitAndClick(approveAuthorizeAccessOcpLocator);
+		await this.driverHelper.waitAndClick(OcpLoginPage.APPROVE_ACCESS_BUTTON);
 	}
 
 	async enterUserNameOpenShift(userName: string): Promise<void> {
 		Logger.debug(`"${userName}"`);
 
-		await this.driverHelper.enterValue(By.id('inputUsername'), userName);
+		await this.driverHelper.enterValue(OcpLoginPage.USERNAME_INPUT, userName);
 	}
 
-	async enterPasswordOpenShift(passw: string): Promise<void> {
+	async enterPasswordOpenShift(userPassword: string): Promise<void> {
 		Logger.debug();
 
-		await this.driverHelper.enterValue(By.id('inputPassword'), passw);
+		await this.driverHelper.enterValue(OcpLoginPage.PASSWORD_INPUT, userPassword);
 	}
 
 	async clickOnLoginButton(): Promise<void> {
 		Logger.debug();
 
-		const loginButtonLocator: By = By.css('button[type=submit]');
-		await this.driverHelper.waitAndClick(loginButtonLocator);
+		await this.driverHelper.waitAndClick(OcpLoginPage.LOGIN_BUTTON);
 	}
 
 	async waitDisappearanceOpenShiftLoginWelcomePage(): Promise<void> {
 		Logger.debug();
 
-		await this.driverHelper.waitDisappearance(By.xpath(OcpLoginPage.LOGIN_PAGE_OPENSHIFT_XPATH));
+		await this.driverHelper.waitDisappearance(OcpLoginPage.LOGIN_PAGE_WELCOME_MESSAGE);
 	}
 }
