@@ -18,17 +18,15 @@ import { e2eContainer } from '../../configs/inversify.config';
 
 @injectable()
 export class OcpImportFromGitPage {
-	private static readonly GIT_URL_INPUT_LOCATOR: By = By.id('form-input-git-url-field');
-	private static readonly SHOW_ADVANCED_GIT_OPTIONS_LINK_LOCATOR: By = By.xpath(
-		'//*[text()="Show advanced Git options"]//ancestor::button'
-	);
-	private static readonly HIDE_ADVANCED_GIT_OPTIONS_LOCATOR: By = By.xpath('//*[text()="Hide advanced Git options"]');
-	private static readonly GIT_REFERENCE_INPUT_LOCATOR: By = By.id('form-input-git-ref-field');
-	private static readonly EDIT_IMPORT_STRATEGY_LINK_LOCATOR: By = By.xpath('//*[text()="Edit Import Strategy"]//ancestor::button');
-	private static readonly BUILDER_IMAGE_STRATEGY_ITEM_LOCATOR: By = By.xpath('//*[text()="Builder Image"]//parent::div//parent::div');
-	private static readonly ADD_LABEL_LINK_LOCATOR: By = By.xpath('//button[text()="Labels"]');
-	private static readonly ADD_LABEL_INPUT_LOCATOR: By = By.id('form-selector-labels-field');
-	private static readonly SUBMIT_BUTTON_LOCATOR: By = By.xpath('//*[@data-test-id="submit-button"]');
+	private static readonly GIT_URL_INPUT: By = By.id('form-input-git-url-field');
+	private static readonly SHOW_ADVANCED_GIT_OPTIONS_LINK: By = By.xpath('//*[text()="Show advanced Git options"]//ancestor::button');
+	private static readonly HIDE_ADVANCED_GIT_OPTIONS: By = By.xpath('//*[text()="Hide advanced Git options"]');
+	private static readonly GIT_REFERENCE_INPUT: By = By.id('form-input-git-ref-field');
+	private static readonly EDIT_IMPORT_STRATEGY_LINK: By = By.xpath('//*[text()="Edit Import Strategy"]//ancestor::button');
+	private static readonly BUILDER_IMAGE_STRATEGY_ITEM: By = By.xpath('//*[text()="Builder Image"]//parent::div//parent::div');
+	private static readonly ADD_LABEL_LINK: By = By.xpath('//button[text()="Labels"]');
+	private static readonly ADD_LABEL_INPUT: By = By.id('form-selector-labels-field');
+	private static readonly SUBMIT_BUTTON: By = By.xpath('//*[@data-test-id="submit-button"]');
 
 	constructor(
 		@inject(CLASSES.DriverHelper)
@@ -38,41 +36,41 @@ export class OcpImportFromGitPage {
 	async enterGitRepoUrl(gitRepoUrl: string): Promise<void> {
 		Logger.debug();
 
-		await this.driverHelper.enterValue(OcpImportFromGitPage.GIT_URL_INPUT_LOCATOR, gitRepoUrl);
+		await this.driverHelper.enterValue(OcpImportFromGitPage.GIT_URL_INPUT, gitRepoUrl);
 	}
 
 	async clickOnAdvancedOptionsButton(): Promise<void> {
 		Logger.debug();
 
-		if (!(await this.driverHelper.isVisible(OcpImportFromGitPage.HIDE_ADVANCED_GIT_OPTIONS_LOCATOR))) {
-			await this.driverHelper.waitAndClick(OcpImportFromGitPage.SHOW_ADVANCED_GIT_OPTIONS_LINK_LOCATOR);
+		if (!(await this.driverHelper.isVisible(OcpImportFromGitPage.HIDE_ADVANCED_GIT_OPTIONS))) {
+			await this.driverHelper.waitAndClick(OcpImportFromGitPage.SHOW_ADVANCED_GIT_OPTIONS_LINK);
 		}
 	}
 
 	async enterGitReference(gitReference: string): Promise<void> {
 		Logger.debug(`"${gitReference}"`);
 
-		await this.driverHelper.enterValue(OcpImportFromGitPage.GIT_REFERENCE_INPUT_LOCATOR, gitReference);
+		await this.driverHelper.enterValue(OcpImportFromGitPage.GIT_REFERENCE_INPUT, gitReference);
 	}
 
 	async selectBuilderImageImportStrategy(): Promise<void> {
 		Logger.debug();
 
-		await this.driverHelper.scrollToAndClick(OcpImportFromGitPage.EDIT_IMPORT_STRATEGY_LINK_LOCATOR);
-		await this.driverHelper.scrollToAndClick(OcpImportFromGitPage.BUILDER_IMAGE_STRATEGY_ITEM_LOCATOR);
+		await this.driverHelper.scrollToAndClick(OcpImportFromGitPage.EDIT_IMPORT_STRATEGY_LINK);
+		await this.driverHelper.scrollToAndClick(OcpImportFromGitPage.BUILDER_IMAGE_STRATEGY_ITEM);
 	}
 
 	async addLabel(label: string): Promise<void> {
 		Logger.debug(`"${label}"`);
 
-		await this.driverHelper.scrollToAndClick(OcpImportFromGitPage.ADD_LABEL_LINK_LOCATOR);
-		await this.driverHelper.scrollToAndEnterValue(OcpImportFromGitPage.ADD_LABEL_INPUT_LOCATOR, label);
+		await this.driverHelper.scrollToAndClick(OcpImportFromGitPage.ADD_LABEL_LINK);
+		await this.driverHelper.scrollToAndEnterValue(OcpImportFromGitPage.ADD_LABEL_INPUT, label);
 	}
 
 	async submitConfiguration(): Promise<OcpApplicationPage> {
 		Logger.debug();
 
-		await this.driverHelper.waitAndClick(OcpImportFromGitPage.SUBMIT_BUTTON_LOCATOR);
+		await this.driverHelper.waitAndClick(OcpImportFromGitPage.SUBMIT_BUTTON);
 		return e2eContainer.get(CLASSES.OcpApplicationPage);
 	}
 
