@@ -13,11 +13,17 @@ import path from 'path';
 import YAML from 'yaml';
 import { expect } from 'chai';
 import { ShellExecutor } from '../../utils/ShellExecutor';
+import { e2eContainer } from '../../configs/inversify.config';
+import { CLASSES } from '../../configs/inversify.types';
 
 suite('Test defining container overrides via attribute.', function (): void {
 	const pathToSampleFile: string = path.resolve('resources/container-overrides.yaml');
 	const workspaceName: string = YAML.parse(fs.readFileSync(pathToSampleFile, 'utf8')).metadata.name;
-	const kubernetesCommandLineToolsExecutor: KubernetesCommandLineToolsExecutor = new KubernetesCommandLineToolsExecutor(workspaceName);
+	const kubernetesCommandLineToolsExecutor: KubernetesCommandLineToolsExecutor = e2eContainer.get(
+		CLASSES.KubernetesCommandLineToolsExecutor
+	);
+	kubernetesCommandLineToolsExecutor.workspaceName = workspaceName;
+	const shellExecutor: ShellExecutor = e2eContainer.get(CLASSES.ShellExecutor);
 
 	suiteSetup('Login into OC client', function (): void {
 		kubernetesCommandLineToolsExecutor.loginToOcp();
@@ -29,7 +35,7 @@ suite('Test defining container overrides via attribute.', function (): void {
 
 	test('Apply container-overrides sample as DevWorkspace with OC client', function (): void {
 		kubernetesCommandLineToolsExecutor.applyYamlConfigurationAsFile(pathToSampleFile);
-		ShellExecutor.wait(5);
+		shellExecutor.wait(5);
 	});
 
 	test('Check that fields are overridden in the Deployment for DevWorkspace', function (): void {
